% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.avails.list}
\alias{accounts.avails.list}
\title{List Avails owned or managed by the partner. See _Authentication and Authorization rules_ and _List methods rules_ for more information about this method.}
\usage{
accounts.avails.list(accountId, pageSize = NULL, pageToken = NULL,
  pphNames = NULL, studioNames = NULL, title = NULL, territories = NULL,
  altId = NULL, videoIds = NULL, altIds = NULL)
}
\arguments{
\item{accountId}{REQUIRED}

\item{pageSize}{See _List methods rules_ for info about this field}

\item{pageToken}{See _List methods rules_ for info about this field}

\item{pphNames}{See _List methods rules_ for info about this field}

\item{studioNames}{See _List methods rules_ for info about this field}

\item{title}{Filter that matches Avails with a `title_internal_alias`, `series_title_internal_alias`, `season_title_internal_alias`, or `episode_title_internal_alias` that contains the given case-insensitive title}

\item{territories}{Filter Avails that match (case-insensitive) any of the given country codes, using the 'ISO 3166-1 alpha-2' format (examples: 'US', 'us', 'Us')}

\item{altId}{Filter Avails that match a case-insensitive, partner-specific custom id}

\item{videoIds}{Filter Avails that match any of the given `video_id`s}

\item{altIds}{Filter Avails that match (case-insensitive) any of the given partner-specific custom ids}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}
