% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.components.list}
\alias{accounts.components.list}
\title{List Components owned or managed by the partner. See _Authentication and Authorization rules_ and _List methods rules_ for more information about this method.}
\usage{
accounts.components.list(accountId, pageSize = NULL, pageToken = NULL,
  pphNames = NULL, studioNames = NULL, titleLevelEidr = NULL,
  editLevelEidr = NULL, status = NULL, customId = NULL,
  inventoryId = NULL, presentationId = NULL, playableSequenceId = NULL,
  elId = NULL, altCutId = NULL, filename = NULL)
}
\arguments{
\item{accountId}{REQUIRED}

\item{pageSize}{See _List methods rules_ for info about this field}

\item{pageToken}{See _List methods rules_ for info about this field}

\item{pphNames}{See _List methods rules_ for info about this field}

\item{studioNames}{See _List methods rules_ for info about this field}

\item{titleLevelEidr}{Filter Components that match a given title-level EIDR}

\item{editLevelEidr}{Filter Components that match a given edit-level EIDR}

\item{status}{Filter Components that match one of the given status}

\item{customId}{Filter Components that match a case-insensitive partner-specific custom id}

\item{inventoryId}{InventoryID available in Common Manifest}

\item{presentationId}{PresentationID available in Common Manifest}

\item{playableSequenceId}{PlayableSequenceID available in Common Manifest}

\item{elId}{Experience ID, as defined by Google}

\item{altCutId}{Filter Components that match a case-insensitive, partner-specific Alternative Cut ID}

\item{filename}{Filter Components that match a case-insensitive substring of the physical name of the delivered file}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}

