% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.experienceLocales.list}
\alias{accounts.experienceLocales.list}
\title{List ExperienceLocales owned or managed by the partner. See _Authentication and Authorization rules_ and _List methods rules_ for more information about this method.}
\usage{
accounts.experienceLocales.list(accountId, pageSize = NULL,
  pageToken = NULL, pphNames = NULL, studioNames = NULL,
  titleLevelEidr = NULL, editLevelEidr = NULL, status = NULL,
  customId = NULL, altCutId = NULL)
}
\arguments{
\item{accountId}{REQUIRED}

\item{pageSize}{See _List methods rules_ for info about this field}

\item{pageToken}{See _List methods rules_ for info about this field}

\item{pphNames}{See _List methods rules_ for info about this field}

\item{studioNames}{See _List methods rules_ for info about this field}

\item{titleLevelEidr}{Filter ExperienceLocales that match a given title-level EIDR}

\item{editLevelEidr}{Filter ExperienceLocales that match a given edit-level EIDR}

\item{status}{Filter ExperienceLocales that match one of the given status}

\item{customId}{Filter ExperienceLocales that match a case-insensitive, partner-specific custom id}

\item{altCutId}{Filter ExperienceLocales that match a case-insensitive, partner-specific Alternative Cut ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}

