% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.orders.get}
\alias{accounts.orders.get}
\title{Get an Order given its id. See _Authentication and Authorization rules_ and _Get methods rules_ for more information about this method.}
\usage{
accounts.orders.get(accountId, orderId)
}
\arguments{
\item{accountId}{REQUIRED}

\item{orderId}{REQUIRED}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}
