% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.orders.list}
\alias{accounts.orders.list}
\title{List Orders owned or managed by the partner. See _Authentication and Authorization rules_ and _List methods rules_ for more information about this method.}
\usage{
accounts.orders.list(accountId, pageSize = NULL, pageToken = NULL,
  pphNames = NULL, studioNames = NULL, name = NULL, status = NULL,
  customId = NULL, videoIds = NULL)
}
\arguments{
\item{accountId}{REQUIRED}

\item{pageSize}{See _List methods rules_ for info about this field}

\item{pageToken}{See _List methods rules_ for info about this field}

\item{pphNames}{See _List methods rules_ for info about this field}

\item{studioNames}{See _List methods rules_ for info about this field}

\item{name}{Filter that matches Orders with a `name`, `show`, `season` or `episode` that contains the given case-insensitive name}

\item{status}{Filter Orders that match one of the given status}

\item{customId}{Filter Orders that match a case-insensitive, partner-specific custom id}

\item{videoIds}{Filter Orders that match any of the given `video_id`s}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}

