% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\name{accounts.storeInfos.list}
\alias{accounts.storeInfos.list}
\title{List StoreInfos owned or managed by the partner. See _Authentication and Authorization rules_ and _List methods rules_ for more information about this method.}
\usage{
accounts.storeInfos.list(accountId, pageSize = NULL, pageToken = NULL,
  pphNames = NULL, studioNames = NULL, videoId = NULL, countries = NULL,
  name = NULL, videoIds = NULL, mids = NULL, seasonIds = NULL)
}
\arguments{
\item{accountId}{REQUIRED}

\item{pageSize}{See _List methods rules_ for info about this field}

\item{pageToken}{See _List methods rules_ for info about this field}

\item{pphNames}{See _List methods rules_ for info about this field}

\item{studioNames}{See _List methods rules_ for info about this field}

\item{videoId}{Filter StoreInfos that match a given `video_id`}

\item{countries}{Filter StoreInfos that match (case-insensitive) any of the given country codes, using the 'ISO 3166-1 alpha-2' format (examples: 'US', 'us', 'Us')}

\item{name}{Filter that matches StoreInfos with a `name` or `show_name` that contains the given case-insensitive name}

\item{videoIds}{Filter StoreInfos that match any of the given `video_id`s}

\item{mids}{Filter StoreInfos that match any of the given `mid`s}

\item{seasonIds}{Filter StoreInfos that match any of the given `season_id`s}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/playmovies_partner.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/playmoviespartner/}{Google Documentation}
}
