% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{Avail}
\alias{Avail}
\title{Google Play Movies Partner API Objects 
Gets the delivery status of titles for Google Play Movies Partners.}
\usage{
Avail(availId = NULL, displayName = NULL, storeLanguage = NULL,
  territory = NULL, workType = NULL, seriesTitleInternalAlias = NULL,
  seasonNumber = NULL, episodeNumber = NULL,
  seasonTitleInternalAlias = NULL, episodeTitleInternalAlias = NULL,
  titleInternalAlias = NULL, licenseType = NULL, formatProfile = NULL,
  start = NULL, end = NULL, priceType = NULL, priceValue = NULL,
  contentId = NULL, productId = NULL, encodeId = NULL,
  seriesAltId = NULL, seasonAltId = NULL, episodeAltId = NULL,
  altId = NULL, suppressionLiftDate = NULL, releaseDate = NULL,
  ratingSystem = NULL, ratingValue = NULL, ratingReason = NULL,
  captionIncluded = NULL, captionExemption = NULL, videoId = NULL,
  pphNames = NULL)
}
\arguments{
\item{availId}{ID internally generated by Google to uniquely identify an Avail}

\item{displayName}{The name of the studio that owns the Edit referred in the Avail}

\item{storeLanguage}{Spoken language of the intended audience}

\item{territory}{ISO 3166-1 alpha-2 country code for the country or territory of this Avail}

\item{workType}{Work type as enumerated in EMA}

\item{seriesTitleInternalAlias}{Title used by involved parties to refer to this series}

\item{seasonNumber}{The number assigned to the season within a series}

\item{episodeNumber}{The number assigned to the episode within a season}

\item{seasonTitleInternalAlias}{Title used by involved parties to refer to this season}

\item{episodeTitleInternalAlias}{OPTIONAL}

\item{titleInternalAlias}{Title used by involved parties to refer to this content}

\item{licenseType}{Type of transaction}

\item{formatProfile}{Indicates the format profile covered by the transaction}

\item{start}{Start of term in YYYY-MM-DD format in the timezone of the country of the Avail}

\item{end}{End of term in YYYY-MM-DD format in the timezone of the country of the Avail}

\item{priceType}{Type of pricing that should be applied to this Avail based on how the partner classify them}

\item{priceValue}{Value to be applied to the pricing type}

\item{contentId}{Title Identifier}

\item{productId}{Edit Identifier}

\item{encodeId}{Manifestation Identifier}

\item{seriesAltId}{Other identifier referring to the series, as defined by partner}

\item{seasonAltId}{Other identifier referring to the season, as defined by partner}

\item{episodeAltId}{Other identifier referring to the episode, as defined by partner}

\item{altId}{Other identifier referring to the Edit, as defined by partner}

\item{suppressionLiftDate}{First date an Edit could be publically announced as becoming available at a specific future date in territory of Avail}

\item{releaseDate}{Release date of the Title in earliest released territory}

\item{ratingSystem}{Rating system applied to the version of title within territory of Avail}

\item{ratingValue}{Value representing the rating}

\item{ratingReason}{Value representing the rating reason}

\item{captionIncluded}{Communicating if caption file will be delivered}

\item{captionExemption}{Communicating an exempt category as defined by FCC regulations}

\item{videoId}{Google-generated ID identifying the video linked to this Avail, once delivered}

\item{pphNames}{Name of the post-production houses that manage the Avail}
}
\value{
Avail object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 19:57:46
filename: /Users/mark/dev/R/autoGoogleAPI/googleplaymoviespartnerv1.auto/R/playmoviespartner_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Avail Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Avail describes the Availability Window of a specific Edit in a given country, which means the period Google is allowed to sell or rent the Edit. Avails are exposed in EMA format Version 1.6b (available at http://www.movielabs.com/md/avails/) Studios can see the Avails for the Titles they own. Post-production houses cannot see any Avails.
}
