% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{Component}
\alias{Component}
\title{Component Object}
\usage{
Component(componentId = NULL, type = NULL, elIds = NULL,
  language = NULL, titleLevelEidrs = NULL, editLevelEidrs = NULL,
  altCutIds = NULL, customIds = NULL, componentDetailType = NULL,
  playableUnitType = NULL, status = NULL, statusDetail = NULL,
  rejectionNote = NULL, approvedTime = NULL, name = NULL,
  processingErrors = NULL, priority = NULL, receivedTime = NULL,
  filename = NULL, studioName = NULL, pphName = NULL,
  normalizedPriority = NULL)
}
\arguments{
\item{componentId}{ID internally generated by Google to uniquely identify the Component}

\item{type}{Type of the Component}

\item{elIds}{IDs internally generated by Google to uniquely identify the ExperienceLocales for which the Component is used}

\item{language}{Language of the component, using the 'BCP 47' format}

\item{titleLevelEidrs}{List of Title-level EIDR IDs}

\item{editLevelEidrs}{List of Edit-level EIDR IDs}

\item{altCutIds}{List of Alternative Cut IDs, sometimes available in lieu of the main Edit-level EIDR IDs}

\item{customIds}{List of custom IDs (defined by the partner) linked to the ExperienceLocale using this Component}

\item{componentDetailType}{Detail about the type of the Component}

\item{playableUnitType}{Type of the playable unit for which the Component is intended}

\item{status}{High-level status of the Component}

\item{statusDetail}{Detailed status of the Component}

\item{rejectionNote}{Notes explaining why a Component has been rejected}

\item{approvedTime}{Timestamp when the Component was approved}

\item{name}{Default Edit name, usually in the language of the country of origin}

\item{processingErrors}{Processing errors during XML file parsing}

\item{priority}{Component priority, as defined by Google}

\item{receivedTime}{Timestamp when the Component was received}

\item{filename}{File name of the Component when delivered}

\item{studioName}{Name of the studio that owns the Component}

\item{pphName}{Name of the post-production house that manages the Component}

\item{normalizedPriority}{A simpler representation of the priority}
}
\value{
Component object
}
\description{
Component Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Component is an element (audio, video, subtitle, artwork, trailer, etc...) that is used in a set of ExperienceLocales. A Component is owned by a Studio and managed either by the Studio itself or by one Post-Production House. Each Component is identified by a `component_id` and its `type`.
}
\seealso{
Other res: \code{\link{Avail}},
  \code{\link{ExperienceLocale}},
  \code{\link{ListAvailsResponse}},
  \code{\link{ListComponentsResponse}},
  \code{\link{ListExperienceLocalesResponse}},
  \code{\link{ListOrdersResponse}},
  \code{\link{ListStoreInfosResponse}},
  \code{\link{Order}}, \code{\link{StoreInfo}}
}

