% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{ExperienceLocale}
\alias{ExperienceLocale}
\title{ExperienceLocale Object}
\usage{
ExperienceLocale(elId = NULL, country = NULL, language = NULL,
  videoId = NULL, trailerId = NULL, titleLevelEidr = NULL,
  editLevelEidr = NULL, altCutId = NULL, customIds = NULL,
  presentationId = NULL, inventoryId = NULL, playableSequenceId = NULL,
  type = NULL, name = NULL, status = NULL, priority = NULL,
  createdTime = NULL, approvedTime = NULL, earliestAvailStartTime = NULL,
  channelId = NULL, studioName = NULL, pphNames = NULL,
  normalizedPriority = NULL)
}
\arguments{
\item{elId}{ID internally generated by Google to uniquely identify a ExperienceLocale}

\item{country}{Country where the ExperienceLocale is available, using the 'ISO 3166-1 alpha-2' format}

\item{language}{Language of the ExperienceLocale, using the 'BCP 47' format}

\item{videoId}{Video ID, as defined by Google, linked to the feature video in the ExperienceLocale}

\item{trailerId}{Trailer ID, as defined by Google, linked to the trailer video in the ExperienceLocale}

\item{titleLevelEidr}{Title-level EIDR ID}

\item{editLevelEidr}{Edit-level EIDR ID}

\item{altCutId}{Alternative Cut ID, sometimes available in lieu of the main Edit-level EIDR ID}

\item{customIds}{List of custom IDs (defined by the partner) linked to this ExperienceLocale}

\item{presentationId}{PresentationID as defined in the EMA specs}

\item{inventoryId}{InventoryID as defined in the EMA specs}

\item{playableSequenceId}{PlayableSequenceID as defined in the EMA specs}

\item{type}{Type of the Edit linked to the ExperienceLocale}

\item{name}{Default Edit name, usually in the language of the country of origin}

\item{status}{High-level status of the ExperienceLocale}

\item{priority}{ExperienceLocale priority, as defined by Google}

\item{createdTime}{Timestamp when the ExperienceLocale was created}

\item{approvedTime}{Timestamp when the ExperienceLocale was approved}

\item{earliestAvailStartTime}{Timestamp of the earliest start date of the Avails linked to this ExperienceLocale}

\item{channelId}{YouTube Channel ID linked to the ExperienceLocale}

\item{studioName}{Name of the studio that owns the ExperienceLocale}

\item{pphNames}{Name of the post-production houses that manage the ExperienceLocale}

\item{normalizedPriority}{A simpler representation of the priority}
}
\value{
ExperienceLocale object
}
\description{
ExperienceLocale Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An ExperienceLocale tracks the fulfillment of a Title in a country using a specific language, when delivered using component-based delivery. For example, a Title in Switzerland might have 3 ExperienceLocales: they both share the same country ('CH'), but each has different languages ('de', 'fr', and 'it'). Each ExperienceLocale is uniquely identified by an `el_id`, which is generated by Google. Externally, an ExperienceLocale can also be identified by partners using its EIDR IDs, AltCutID or `custom_id` (when provided).
}
\seealso{
Other res: \code{\link{Avail}}, \code{\link{Component}},
  \code{\link{ListAvailsResponse}},
  \code{\link{ListComponentsResponse}},
  \code{\link{ListExperienceLocalesResponse}},
  \code{\link{ListOrdersResponse}},
  \code{\link{ListStoreInfosResponse}},
  \code{\link{Order}}, \code{\link{StoreInfo}}
}

