% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{ListAvailsResponse}
\alias{ListAvailsResponse}
\title{ListAvailsResponse Object}
\usage{
ListAvailsResponse(avails = NULL, nextPageToken = NULL, totalSize = NULL)
}
\arguments{
\item{avails}{List of Avails that match the request criteria}

\item{nextPageToken}{See _List methods rules_ for info about this field}

\item{totalSize}{See _List methods rules_ for more information about this field}
}
\value{
ListAvailsResponse object
}
\description{
ListAvailsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to the 'ListAvails' method.
}
