% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{ListOrdersResponse}
\alias{ListOrdersResponse}
\title{ListOrdersResponse Object}
\usage{
ListOrdersResponse(orders = NULL, nextPageToken = NULL, totalSize = NULL)
}
\arguments{
\item{orders}{List of Orders that match the request criteria}

\item{nextPageToken}{See _List methods rules_ for info about this field}

\item{totalSize}{See _List methods rules_ for more information about this field}
}
\value{
ListOrdersResponse object
}
\description{
ListOrdersResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to the 'ListOrders' method.
}
