% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{ListStoreInfosResponse}
\alias{ListStoreInfosResponse}
\title{ListStoreInfosResponse Object}
\usage{
ListStoreInfosResponse(storeInfos = NULL, nextPageToken = NULL,
  totalSize = NULL)
}
\arguments{
\item{storeInfos}{List of StoreInfos that match the request criteria}

\item{nextPageToken}{See 'List methods rules' for info about this field}

\item{totalSize}{See _List methods rules_ for more information about this field}
}
\value{
ListStoreInfosResponse object
}
\description{
ListStoreInfosResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response to the 'ListStoreInfos' method.
}
