% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{Order}
\alias{Order}
\title{Order Object}
\usage{
Order(orderId = NULL, customId = NULL, videoId = NULL, countries = NULL,
  type = NULL, name = NULL, episodeName = NULL, seasonName = NULL,
  showName = NULL, status = NULL, statusDetail = NULL,
  rejectionNote = NULL, orderedTime = NULL, approvedTime = NULL,
  receivedTime = NULL, earliestAvailStartTime = NULL, priority = NULL,
  legacyPriority = NULL, channelId = NULL, channelName = NULL,
  studioName = NULL, pphName = NULL, normalizedPriority = NULL)
}
\arguments{
\item{orderId}{ID internally generated by Google to uniquely identify an Order}

\item{customId}{ID that can be used to externally identify an Order}

\item{videoId}{Google-generated ID identifying the video linked to this Order, once delivered}

\item{countries}{Countries where the Order is available, using the 'ISO 3166-1 alpha-2' format (example: 'US')}

\item{type}{Type of the Edit linked to the Order}

\item{name}{Default Edit name, usually in the language of the country of origin}

\item{episodeName}{Default Episode name, usually in the language of the country of origin}

\item{seasonName}{Default Season name, usually in the language of the country of origin}

\item{showName}{Default Show name, usually in the language of the country of origin}

\item{status}{High-level status of the order}

\item{statusDetail}{Detailed status of the order}

\item{rejectionNote}{Field explaining why an Order has been rejected}

\item{orderedTime}{Timestamp when the Order was created}

\item{approvedTime}{Timestamp when the Order was approved}

\item{receivedTime}{Timestamp when the Order was fulfilled}

\item{earliestAvailStartTime}{Timestamp of the earliest start date of the Avails linked to this Order}

\item{priority}{Order priority, as defined by Google}

\item{legacyPriority}{Legacy Order priority, as defined by Google}

\item{channelId}{YouTube Channel ID that should be used to fulfill the Order}

\item{channelName}{YouTube Channel Name that should be used to fulfill the Order}

\item{studioName}{Name of the studio that owns the Edit ordered}

\item{pphName}{Name of the post-production house that manages the Edit ordered}

\item{normalizedPriority}{A simpler representation of the priority}
}
\value{
Order object
}
\description{
Order Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An Order tracks the fulfillment of an Edit when delivered using the legacy, non-component-based delivery. Each Order is uniquely identified by an `order_id`, which is generated by Google. Externally, Orders can also be identified by partners using its `custom_id` (when provided).
}
\seealso{
Other res: \code{\link{Avail}}, \code{\link{Component}},
  \code{\link{ExperienceLocale}},
  \code{\link{ListAvailsResponse}},
  \code{\link{ListComponentsResponse}},
  \code{\link{ListExperienceLocalesResponse}},
  \code{\link{ListOrdersResponse}},
  \code{\link{ListStoreInfosResponse}},
  \code{\link{StoreInfo}}
}

