% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_functions.R
\docType{package}
\name{playmoviespartner_googleAuthR}
\alias{playmoviespartner_googleAuthR}
\alias{playmoviespartner_googleAuthR-package}
\title{Google Play Movies Partner API
Gets the delivery status of titles for Google Play Movies Partners.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:21:31
filename: /Users/mark/dev/R/autoGoogleAPI/googleplaymoviespartnerv1.auto/R/playmoviespartner_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/playmovies_partner.readonly
}
}

