% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playmoviespartner_objects.R
\name{StoreInfo}
\alias{StoreInfo}
\title{StoreInfo Object}
\usage{
StoreInfo(videoId = NULL, seasonId = NULL, showId = NULL,
  country = NULL, liveTime = NULL, type = NULL, name = NULL,
  titleLevelEidr = NULL, editLevelEidr = NULL, seasonName = NULL,
  showName = NULL, seasonNumber = NULL, episodeNumber = NULL,
  hasSdOffer = NULL, hasHdOffer = NULL, hasVodOffer = NULL,
  hasEstOffer = NULL, hasAudio51 = NULL, audioTracks = NULL,
  subtitles = NULL, hasInfoCards = NULL, mid = NULL, trailerId = NULL,
  studioName = NULL, pphNames = NULL)
}
\arguments{
\item{videoId}{Google-generated ID identifying the video linked to the Edit}

\item{seasonId}{Google-generated ID identifying the season linked to the Edit}

\item{showId}{Google-generated ID identifying the show linked to the Edit}

\item{country}{Country where Edit is available in ISO 3166-1 alpha-2 country code}

\item{liveTime}{Timestamp when the Edit went live on the Store}

\item{type}{Edit type, like Movie, Episode or Season}

\item{name}{Default Edit name, usually in the language of the country of origin}

\item{titleLevelEidr}{Title-level EIDR ID}

\item{editLevelEidr}{Edit-level EIDR ID}

\item{seasonName}{Default Season name, usually in the language of the country of origin}

\item{showName}{Default Show name, usually in the language of the country of origin}

\item{seasonNumber}{The number assigned to the season within a show}

\item{episodeNumber}{The number assigned to the episode within a season}

\item{hasSdOffer}{Whether the Edit has a SD offer}

\item{hasHdOffer}{Whether the Edit has a HD offer}

\item{hasVodOffer}{Whether the Edit has a VOD offer}

\item{hasEstOffer}{Whether the Edit has a EST offer}

\item{hasAudio51}{Whether the Edit has a 5}

\item{audioTracks}{Audio tracks available for this Edit}

\item{subtitles}{Subtitles available for this Edit}

\item{hasInfoCards}{Whether the Edit has info cards}

\item{mid}{Knowledge Graph ID associated to this Edit, if available}

\item{trailerId}{Google-generated ID identifying the trailer linked to the Edit}

\item{studioName}{Name of the studio that owns the Edit ordered}

\item{pphNames}{Name of the post-production houses that manage the Edit}
}
\value{
StoreInfo object
}
\description{
StoreInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a playable sequence (video) associated with an Edit and available at the Google Play Store. Internally, each StoreInfo is uniquely identified by a `video_id` and `country`. Externally, Title-level EIDR or Edit-level EIDR, if provided, can also be used to identify a specific title or edit in a country.
}
